+++ R/PipeOpLearner.R
@@ -62,7 +62,17 @@
 #'   [`Learner`][mlr3::Learner] that is being wrapped. Read-only.
 #' * `learner_model` :: [`Learner`][mlr3::Learner]\cr
 #'   [`Learner`][mlr3::Learner] that is being wrapped. This learner contains the model if the `PipeOp` is trained. Read-only.
-#'
+#' * `validate` :: `"predefined"` or `NULL`\cr
+#'    This field can only be set for `Learner`s that have the `"validation"` property.
+#'    Setting the field to `"predefined"` means that the wrapped `Learner` will use the internal validation task,
+#'    otherwise it will be ignored.
+#'    Note that specifying *how* the validation data is created is possible via the `$validate` field of the [`GraphLearner`].
+#'    For each `PipeOp` it is then only possible to either use it (`"predefined"`) or not use it (`NULL`).
+#'    Also see [`set_validate.GraphLearner`] for more information.
+#' * `internal_tuned_values` :: named `list()` or `NULL`\cr
+#'    The internally tuned values if the wrapped `Learner` supports internal tuning, `NULL` otherwise.
+#' * `internal_valid_scores` :: named `list()` or `NULL`\cr
+#'    The internal validation scores if the wrapped `Learner` supports internal validation, `NULL` otherwise.
 #' @section Methods:
 #' Methods inherited from [`PipeOp`].
 #'
@@ -87,18 +97,48 @@ PipeOpLearner = R6Class("PipeOpLearner", inherit = PipeOp,
     initialize = function(learner, id = NULL, param_vals = list()) {
       private$.learner = as_learner(learner, clone = TRUE)
       id = id %??% private$.learner$id
+      if (!test_po_validate(get0("validate", private$.learner))) {
+        stopf(
+          "Validate field of PipeOp '%s' must either be NULL or 'predefined'.\nTo configure how the validation data is created, set the $validate field of the GraphLearner, e.g. using set_validate().",  # nolint
+          id
+        )
+      }
       # FIXME: can be changed when mlr-org/mlr3#470 has an answer
       type = private$.learner$task_type
       task_type = mlr_reflections$task_types[type, mult = "first"]$task
       out_type = mlr_reflections$task_types[type, mult = "first"]$prediction
+      properties = c("validation", "internal_tuning")
+      properties = properties[properties %in% private$.learner$properties]
       super$initialize(id, param_set = alist(private$.learner$param_set), param_vals = param_vals,
         input = data.table(name = "input", train = task_type, predict = task_type),
         output = data.table(name = "output", train = "NULL", predict = out_type),
-        tags = "learner", packages = learner$packages
+        tags = "learner", packages = learner$packages, properties = properties
       )
     }
   ),
   active = list(
+    internal_tuned_values = function(rhs) {
+      assert_ro_binding(rhs)
+      if ("internal_tuning" %nin% self$properties) return(NULL)
+      self$learner_model$internal_tuned_values
+    },
+    internal_valid_scores = function(rhs) {
+      assert_ro_binding(rhs)
+      if ("validation" %nin% self$properties) return(NULL)
+      self$learner_model$internal_valid_scores
+    },
+    validate = function(rhs) {
+      if ("validation" %nin% self$properties) {
+        if (!missing(rhs)) {
+          stopf("PipeOp '%s' does not support validation, because the wrapped Learner doesn't.", self$id)
+        }
+        return(NULL)
+      }
+      if (!missing(rhs)) {
+        private$.learner$validate = assert_po_validate(rhs)
+      }
+      private$.learner$validate
+    },
     id = function(val) {
       if (!missing(val)) {
         private$.id = val
@@ -112,6 +152,14 @@ PipeOpLearner = R6Class("PipeOpLearner", inherit = PipeOp,
       if (!missing(val)) {
         if (!identical(val, private$.learner)) {
           stop("$learner is read-only.")
+        }
+        validate = get0("validate", private$.learner)
+        if (!test_po_validate(validate)) {
+          warningf(paste(sep = "\n",
+            "PipeOpLearner '%s' has its validate field set to a value that is neither NULL nor 'predefined'.",
+            "This will likely lead to unexpected behaviour.",
+            "Configure the $validate field of the GraphLearner to define how the validation data is created."
+            ), self$id)
         }
       }
       private$.learner
@@ -157,4 +205,30 @@ PipeOpLearner = R6Class("PipeOpLearner", inherit = PipeOp,
   )
 )
 
-mlr_pipeops$add("learner", PipeOpLearner, list(R6Class("Learner", public = list(id = "learner", task_type = "classif", param_set = ps(), packages = "mlr3pipelines"))$new()))
+mlr_pipeops$add("learner", PipeOpLearner, list(R6Class("Learner", public = list(properties = character(0), id = "learner", task_type = "classif", param_set = ps(), packages = "mlr3pipelines"))$new())) # nolint
+
+#' @export
+set_validate.PipeOpLearner = function(learner, validate, ...) {
+  assert_po_validate(validate)
+  on.exit({
+    # don't use set_validate(), as it also does not work in general and might fail again
+    learner$validate = prev_validate
+  })
+  prev_validate = learner$validate
+  withCallingHandlers({
+    set_validate(learner$learner, validate = validate, ...)
+  }, error = function(e) {
+    e$message = sprintf(paste0(
+      "Failed to set validate for Learner '%s':\n%s\n",
+      "Trying to heuristically reset validation to its previous state, please check the results"), learner$id, e$message)
+    stop(e)
+  }, warning = function(w) {
+    w$message = sprintf(paste0(
+      "Failed to set validate for PipeOp '%s':\n%s\n",
+      "Trying to heuristically reset validation to its previous state, please check the results"), learner$id, w$message)
+    warning(w)
+    invokeRestart("muffleWarning")
+  })
+  on.exit()
+  invisible(learner)
+}